+++ extensions/hidemyemail/src/api/hide-my-email.ts
@@ -1,10 +1,14 @@
-import { AxiosRequestConfig } from "axios";
+import axios, { AxiosRequestConfig } from "axios";
 import { iCloudService, iCloudSession } from "./connect";
 import { iCloudAPIResponseError } from "./errors";
 
-export interface HideMyEmail {
+interface IHideMyEmail {
+  originAppName?: string;
+  appIconUrl?: string;
+  appBundleId?: string;
   origin: string;
   anonymousId: string;
+  domain: string;
   forwardToEmail: string;
   hme: string;
   label: string;
@@ -12,6 +16,70 @@ export interface HideMyEmail {
   createTimestamp: number;
   isActive: boolean;
   recipientMailId: string;
+}
+
+export class HideMyEmail {
+  id: string;
+  address: string;
+  label: string;
+  note: string;
+  forwardToEmail: string;
+  origin: string;
+  domain: string;
+  appID?: string;
+  appIconURL?: string;
+  createTimestamp: number;
+  isActive: boolean;
+
+  constructor(hme: IHideMyEmail) {
+    this.id = hme.anonymousId;
+    this.address = hme.hme;
+    this.label = hme.label;
+    this.note = hme.note;
+    this.domain = hme.domain;
+    this.forwardToEmail = hme.forwardToEmail;
+    this.createTimestamp = hme.createTimestamp;
+    this.isActive = hme.isActive;
+
+    switch (hme.origin) {
+      case "ON_DEMAND":
+        this.origin = "Hide My Email";
+        this.appID = "default";
+        break;
+      case "SAFARI":
+        this.origin = "Safari";
+        this.appID = "com.apple.mobilesafari";
+        break;
+      case "MAIL":
+        this.origin = "Mail";
+        this.appID = "com.apple.mobilemail";
+        break;
+      case "IN_APP":
+        this.origin = hme.originAppName!;
+        if (hme.appBundleId) {
+          this.appID = hme.appBundleId;
+        }
+        break;
+      default:
+        this.origin = hme.origin;
+    }
+  }
+}
+
+export async function getIcon(hme: HideMyEmail) {
+  try {
+    const response = await axios.get(`https://itunes.apple.com/search?term=${hme.origin}&limit=3&entity=software`);
+    console.log(hme.appID);
+    console.log(response.data.results);
+    for (const entry of response.data.results) {
+      if (entry.bundleId === hme.appID) {
+        console.log("FOUND!", entry.artworkUrl512);
+        return entry.artworkUrl512;
+      }
+    }
+  } catch (error) {
+    return "";
+  }
 }
 
 export interface MetaData {
@@ -46,12 +114,19 @@ export class HideMyEmailService {
       console.log("Failed to fetch addresses: ", response.data);
       throw new iCloudAPIResponseError(response.data.error.errorMessage, response.data.error.errorCode);
     }
-    return response.data.result?.hmeEmails;
+
+    if (!response.data.result?.hmeEmails) return [];
+    const rawEmails = response.data.result?.hmeEmails;
+    const emails = [];
+    for (const hme of rawEmails) {
+      emails.push(new HideMyEmail(hme));
+    }
+    return emails;
   }
 
   async toggleActive(email: HideMyEmail, toggle: AllowedValues, axiosConfig: AxiosRequestConfig = {}) {
     const endPoint = `${this.emailEndpointUpdate}/` + toggle;
-    const data = { anonymousId: email.anonymousId };
+    const data = { anonymousId: email.id };
 
     const response = await this.session.request("post", endPoint, { data, ...axiosConfig }, false);
     if (response.data.success === false) {
@@ -62,7 +137,7 @@ export class HideMyEmailService {
 
   async updateMetaData(email: HideMyEmail, newMetaData: MetaData, axiosConfig: AxiosRequestConfig = {}) {
     const endPoint = `${this.emailEndpointUpdate}/updateMetaData`;
-    const data = { anonymousId: email.anonymousId, ...newMetaData };
+    const data = { anonymousId: email.id, ...newMetaData };
     const response = await this.session.request("post", endPoint, { data, ...axiosConfig }, false);
 
     if (response.data.success === false) {
@@ -75,7 +150,7 @@ export class HideMyEmailService {
     if (email.isActive) await this.toggleActive(email, "deactivate");
 
     const endPoint = `${this.emailEndpointUpdate}/delete`;
-    const data = { anonymousId: email.anonymousId };
+    const data = { anonymousId: email.id };
 
     const response = await this.session.request("post", endPoint, { data, ...axiosConfig }, false);
     if (response.data.success === false) {
@@ -95,8 +170,8 @@ export class HideMyEmailService {
       throw new iCloudAPIResponseError(response.data.error.errorMessage, response.data.error?.errorCode);
     }
 
-    const address = response.data.result["hme"];
-    return address;
+    const hme = response.data.result["hme"];
+    return hme;
   }
 
   async addAddress(address: string, metaData: MetaData, axiosConfig: AxiosRequestConfig = {}) {