+++ extensions/hidemyemail/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Improved user experience through caching and optimistic updates
 - Increased timeout limit for network requests to reduce failed requests
 - Added icons for some actions
+- Handled situation where session expires while using the extension
+- Added more information in metadata section
 
 ## [Update] - 2024-12-20